v0.2.4

+++ lib/neptuno/docker/down.rb
@@ -6,6 +6,7 @@ module Neptuno
     class Down < Neptuno::CLI::Base
       desc "Docker: Stop docker containers for current project"
 
+      option :tmux, type: :boolean, default: false, desc: "Kill tmux session"
       option :all, type: :boolean, default: false, desc: "Run on all services"
       argument :services, type: :array, required: false, desc: "Optional list of services"
 
@@ -13,7 +14,7 @@ module Neptuno
         command_services_to("go down", all: options.fetch(:all), services_as_args: services) do |services|
           system("cd #{neptuno_path} && docker-compose stop -t 0 #{services.join(" ")}")
           system("cd #{neptuno_path} && docker-compose rm -f #{services.join(" ")}")
-          system("cd #{neptuno_path} && tmux kill-session -t neptuno")
+          system("cd #{neptuno_path} && tmux kill-session -t neptuno") if options.fetch(:tmux)
         end
       end
     end
